small fixes

+++ maintenance/includes/tasks/import06/import06.class.php
@@ -512,7 +512,7 @@ class import06 extends task {
 			$logs = array();
 			list($d,$m,$y) = explode('.',$this->get('logs_date', '1.1.1970'));
 			$date = mktime(0,0,0,$m,$d,$y);
-			$sql = "SELECT * FROM ".$this->old[1]."logs WHERE log_date > ".$date." LIMIT 5000;";
+			$sql = "SELECT l.*, u.username FROM ".$this->old[1]."logs l LEFT JOIN ".$this->old[1]."users u ON l.admin_id=u.user_id WHERE log_date > ".$date." LIMIT 5000;";
 			$result = $this->old[0]->query($sql);
 			include($this->root_path.'maintenance/includes/tasks/import06/plugin_logs.php');
 			while ( $row = $this->old[0]->fetch_record($result) ) {
@@ -534,7 +534,11 @@ class import06 extends task {
 				$current_log['log_tag'] = strtolower($to_replace[1]);
 				$current_log['log_plugin'] = (in_array($row['log_type'], array_keys($plugin_logs))) ? $this->new[0]->escape($plugin_logs[$row['log_type']]) : 'core';
 				$current_log['log_flag'] = 1;
-				$current_log['user_id'] = $this->new[0]->escape($row['admin_id']);
+				if(isset($this->step_data['replace_users'][$row['admin_id']])) {
+					$current_log['user_id'] = $this->step_data['replace_users'][$row['admin_id']];
+				} else $current_log['user_id'] = -1;
+				$current_log['username'] = ($row['username'] == "") ? "Unknown" : $this->new[0]->escape($row['username']);
+				
 				if(in_array($current_log['log_plugin'], $copy)) $logs[$row['log_id']] = $current_log;
 			}
 			$this->old[0]->free_result($result);
@@ -993,49 +997,51 @@ class import06 extends task {
 		}
 		$this->old[0]->free_result($result);
 		unset($this->step_data['import_data']['raids']);
-		$fields = "`".implode('`, `', array_keys(current($raids)))."`";
-		$this->new[0]->query("TRUNCATE ".$this->new[1]."raids;");
-		$this->new[0]->query("TRUNCATE ".$this->new[1]."raid_attendees;");
-		$sql = "INSERT INTO ".$this->new[1]."raids (".$fields.") VALUES ";
-		$sqls = array();
-		$ra_ids = array();
-		foreach($raids as $raid_i => $raid) {
-			$sqls[] = "('".implode("', '", $raid)."')";
-			$ra_ids[] = $raid_i;
-		}
-		$this->new[0]->query($sql.implode(', ', $sqls).';');
-		unset($raids);
-		$raid_attendees = array();
-		$sql = "SELECT raid_id, member_name FROM ".$this->old[1]."raid_attendees WHERE raid_id IN (".implode(',',$ra_ids).");";
-		$result = $this->old[0]->query($sql);
-		while ( $row = $this->old[0]->fetch_record($result) ) {
-			$raid_attendees[$row['raid_id']][] = $row['member_name'];
-		}
-		$this->old[0]->free_result($result);
-		$members = array();
-		$result = $this->new[0]->query("SELECT member_name, member_id FROM ".$this->new[1]."members;");
-		while ( $row = $this->new[0]->fetch_record($result) ) {
-			$members[$row['member_name']] = $row['member_id'];
-		}
-		$this->new[0]->free_result($result);
-		$raid_atts = array();
-		foreach($raid_attendees as $raid_id => $atts) {
-			foreach($atts as $name) {
-				if($members[$name]) {
-					$raid_atts[$raid_id][] = $members[$name];
+		if (count($raids)){
+			$fields = "`".implode('`, `', array_keys(current($raids)))."`";
+			$this->new[0]->query("TRUNCATE ".$this->new[1]."raids;");
+			$this->new[0]->query("TRUNCATE ".$this->new[1]."raid_attendees;");
+			$sql = "INSERT INTO ".$this->new[1]."raids (".$fields.") VALUES ";
+			$sqls = array();
+			$ra_ids = array();
+			foreach($raids as $raid_i => $raid) {
+				$sqls[] = "('".implode("', '", $raid)."')";
+				$ra_ids[] = $raid_i;
+			}
+			$this->new[0]->query($sql.implode(', ', $sqls).';');
+			unset($raids);
+			$raid_attendees = array();
+			$sql = "SELECT raid_id, member_name FROM ".$this->old[1]."raid_attendees WHERE raid_id IN (".implode(',',$ra_ids).");";
+			$result = $this->old[0]->query($sql);
+			while ( $row = $this->old[0]->fetch_record($result) ) {
+				$raid_attendees[$row['raid_id']][] = $row['member_name'];
+			}
+			$this->old[0]->free_result($result);
+			$members = array();
+			$result = $this->new[0]->query("SELECT member_name, member_id FROM ".$this->new[1]."members;");
+			while ( $row = $this->new[0]->fetch_record($result) ) {
+				$members[$row['member_name']] = $row['member_id'];
+			}
+			$this->new[0]->free_result($result);
+			$raid_atts = array();
+			foreach($raid_attendees as $raid_id => $atts) {
+				foreach($atts as $name) {
+					if($members[$name]) {
+						$raid_atts[$raid_id][] = $members[$name];
+					}
 				}
 			}
-		}
-		$sql = "INSERT INTO ".$this->new[1]."raid_attendees (raid_id, member_id) VALUES ";
-		$sqls = array();
-		foreach($raid_atts as $raid_id => $atts) {
-			foreach($atts as $member_id) {
-				$sqls[] = "('".$raid_id."', '".$member_id."')";
+			$sql = "INSERT INTO ".$this->new[1]."raid_attendees (raid_id, member_id) VALUES ";
+			$sqls = array();
+			foreach($raid_atts as $raid_id => $atts) {
+				foreach($atts as $member_id) {
+					$sqls[] = "('".$raid_id."', '".$member_id."')";
+				}
 			}
+			$this->new[0]->query($sql.implode(', ', $sqls).';');
+			unset($raid_atts);
+			unset($raid_attendees);
 		}
-		$this->new[0]->query($sql.implode(', ', $sqls).';');
-		unset($raid_atts);
-		unset($raid_attendees);
 		return true;
 	}
